Important bug fix in pulsegen: earlier input message was of type int: hence output(float) from one pulsegen when fed into another's input, it would behave erratically. Now input message is of type double.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@2409 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 61021b5754a4811f8e982e8598e7ee3dd9a0d651

+++ device/PulseGen.cpp
@@ -304,7 +304,7 @@ void PulseGen::setPulseDelay(const Conn* c, int index, double delay)
     index == 0? obj->firstDelay_ = delay: obj->secondDelay_ = delay;
 }
 
-void PulseGen::inputFunc(const Conn* c, int value)
+void PulseGen::inputFunc(const Conn* c, double value)
 {
     PulseGen* obj = static_cast<PulseGen*> (c->data());
     ASSERT( obj != NULL, "PulseGen::inputFunc(const Conn*, int) - target data pointer is NULL.");
